refactor: move http url override to network common module
