+++ embrace-android-instrumentation-huc/src/test/kotlin/io/embrace/android/embracesdk/instrumentation/huc/EmbraceUrlConnectionDelegateTest.kt
@@ -4,7 +4,7 @@ import io.embrace.android.embracesdk.fakes.FakeEmbraceInternalInterface
 import io.embrace.android.embracesdk.instrumentation.huc.EmbraceUrlConnectionDelegate.CONTENT_ENCODING
 import io.embrace.android.embracesdk.instrumentation.huc.EmbraceUrlConnectionDelegate.CONTENT_LENGTH
 import io.embrace.android.embracesdk.internal.config.behavior.NetworkSpanForwardingBehaviorImpl.Companion.TRACEPARENT_HEADER_NAME
-import io.embrace.android.embracesdk.internal.network.logging.PATH_OVERRIDE
+import io.embrace.android.embracesdk.internal.instrumentation.network.PATH_OVERRIDE
 import io.embrace.android.embracesdk.network.http.HttpMethod
 import io.mockk.every
 import io.mockk.mockk
