+++ embrace-android-instrumentation-network-common/src/main/kotlin/io/embrace/android/embracesdk/internal/instrumentation/network/HttpPathOverrideRequest.kt
@@ -1,4 +1,4 @@
-package io.embrace.android.embracesdk.internal.network.logging
+package io.embrace.android.embracesdk.internal.instrumentation.network
 
 interface HttpPathOverrideRequest {
     fun getHeaderByName(name: String): String?
